Add example of the beat classifier
Added an overlap feature to the PCT. The class only works for the
initial run. Outside of the initial run values become wonky.

Cleaned up the adap threshold file. Renamed the beat classifier function
to beat finder, then created a new classifier function that compares the
values for the finder for two frequency bands.

Added an example of how to use the classifier and polynomialCT class.